chore: move arouund fake chunk generator
